requirements.txt
pandas
numpy
requests
python-dotenv
websocket-client
xgboost
scikit-learn
pytest
setuptools
wheel

